Refacttored skolemization logic
... into a new trait "Skolemization".
